Initial reflective serialization

+++ CuddleKit/Format/Int64Formatter.cs
@@ -4,7 +4,7 @@ using CuddleKit.Serialization;
 
 namespace CuddleKit.Format
 {
-	public sealed class Int64Formatter : Formatter<int>
+	public sealed class Int64Formatter : Formatter<long>
 	{
 		private const int BufferLength = 20; // 19 for the longest input: 9,223,372,036,854,775,807
 
@@ -16,10 +16,10 @@ namespace CuddleKit.Format
 			: base(DataType.Integer, "i64", flags, BufferLength) =>
 			_provider = provider;
 
-		protected override bool TryImport(ReadOnlySpan<char> source, out int value) =>
-			int.TryParse(source, NumberStyles.Integer, _provider, out value);
+		protected override bool TryImport(ReadOnlySpan<char> source, out long value) =>
+			long.TryParse(source, NumberStyles.Integer, _provider, out value);
 
-		protected override bool TryExport(int value, Span<char> destination, out int exportedLength) =>
+		protected override bool TryExport(long value, Span<char> destination, out int exportedLength) =>
 			value.TryFormat(destination, out exportedLength, default, _provider);
 	}
 }
